chore: rename directory, template -> market_price
