+++ market_price/base/miner.py
@@ -23,8 +23,8 @@ import traceback
 
 import bittensor as bt
 
-from template.base.neuron import BaseNeuron
-from template.utils.config import add_miner_args
+from market_price.base.neuron import BaseNeuron
+from market_price.utils.config import add_miner_args
 
 from typing import Union
