+++ market_price/base/neuron.py
@@ -23,10 +23,10 @@ import bittensor as bt
 from abc import ABC, abstractmethod
 
 # Sync calls set weights and also resyncs the metagraph.
-from template.utils.config import check_config, add_args, config
-from template.utils.misc import ttl_get_block
-from template import __spec_version__ as spec_version
-from template.mock import MockSubtensor, MockMetagraph
+from market_price.utils.config import check_config, add_args, config
+from market_price.utils.misc import ttl_get_block
+from market_price import __spec_version__ as spec_version
+from market_price.mock import MockSubtensor, MockMetagraph
 
 
 class BaseNeuron(ABC):
