+++ market_price/base/validator.py
@@ -28,13 +28,13 @@ import bittensor as bt
 from typing import List, Union
 from traceback import print_exception
 
-from template.base.neuron import BaseNeuron
-from template.base.utils.weight_utils import (
+from market_price.base.neuron import BaseNeuron
+from market_price.base.utils.weight_utils import (
     process_weights_for_netuid,
     convert_weights_and_uids_for_emit,
 )  # TODO: Replace when bittensor switches to numpy
-from template.mock import MockDendrite
-from template.utils.config import add_validator_args
+from market_price.mock import MockDendrite
+from market_price.utils.config import add_validator_args
 
 
 class BaseValidatorNeuron(BaseNeuron):
