+++ market_price/utils/uids.py
