+++ market_price/validator/forward.py
@@ -20,9 +20,9 @@
 import time
 import bittensor as bt
 
-from template.protocol import Dummy
-from template.validator.reward import get_rewards
-from template.utils.uids import get_random_uids
+from market_price.protocol import MarketPriceSynapse
+from market_price.validator.reward import get_rewards
+from market_price.utils.uids import get_random_uids
 
 
 async def forward(self):
@@ -44,7 +44,7 @@ async def forward(self):
         # Send the query to selected miner axons in the network.
         axons=[self.metagraph.axons[uid] for uid in miner_uids],
         # Construct a dummy query. This simply contains a single integer.
-        synapse=Dummy(timestamp="2024-10-11T22:26:00+01:00"),
+        synapse=MarketPriceSynapse(timestamp="2024-10-11T22:26:00+01:00"),
         # All responses have the deserialize function called on them before returning.
         # You are encouraged to define your own deserialization function.
         deserialize=True,
