+++ tests/test_template_validator.py
@@ -23,10 +23,10 @@ import bittensor as bt
 import torch
 
 from neurons.validator import Validator
-from template.base.validator import BaseValidatorNeuron
-from template.protocol import Dummy
-from template.utils.uids import get_random_uids
-from template.validator.reward import get_rewards
+from market_price.base.validator import BaseValidatorNeuron
+from market_price.protocol import MarketPriceSynapse
+from market_price.utils.uids import get_random_uids
+from market_price.validator.reward import get_rewards
 
 
 class TemplateValidatorNeuronTestCase(unittest.TestCase):
